- Added a test-gateway Justfile recipe that invokes cargo test -p the_block --lib --features gateway web::gateway::tests::, giving CI a direct entry point for the feature-gated HTTP suite.
- Gateway read acknowledgement handling now resolves providers through pipeline::provider_for_manifest, attaches campaign metadata, and verifies badge-targeted matches via new tests.
- Storage pipeline gained deterministic provider selection helpers plus override hooks for manifests, blobs, and WASM assets to keep tests hermetic.
- Updated storage pipeline documentation to explain manifest-based provider resolution and the new override utilities used by gateway tests.
- RPC state now carries an optional MarketplaceHandle, and the ad-market module returns explicit status or error codes when inventory, distribution, or registration endpoints run without backing storage.
- Introduced node/tests/ad_market_rpc.rs to drive the RPC server in-process, covering happy paths, duplicates, invalid payloads, and concurrent registrations against the sled-backed marketplace.
- Clippy cleanups swept through first-party async, base64, and X25519 helpers by adding Display impls, restructuring loops, and sharing clamp utilities.
- Monitoring dashboards now render read and advertising payout last-seen and staleness panels with role legends, with tests asserting every Grafana template contains the new panels.
- Alert validator fixtures gained payout last-seen samples and evaluation helpers so the Explorer staleness alerts stay audited in CI.
- Progress and roadmap logs capture the October 25 reviews detailing the lint cleanup, gateway recipe, ad-market harness, and monitoring visibility commitments.
- Gateway documentation now outlines manifest-driven provider selection, badge context, override hooks, and signature coverage to align operator workflows with the updated implementation.

+++ crates/rand_core/src/lib.rs
@@ -89,7 +89,7 @@ impl OsRng {
         let since_epoch = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .map_err(|_| Error::new(ErrorKind::Unavailable))?;
-        let nanos = since_epoch.as_nanos() as u128;
+        let nanos = since_epoch.as_nanos();
         Ok(Self {
             state: nanos ^ 0x9e37_79b9_7f4a_7c15_ffff,
         })
